+++ tools/dev_riderbark.lua
@@ -22,28 +22,12 @@
 local SPEAK = nil
 
 local LINES = {
-	{ "q_theresa_village_m_henryTraining", "Ha! / Take that!" },
-	{ "q_disguise_ondrejDialog_dealDiscussion", "What!?" },
-	{ "q_superstition_startPhase3", "Alright. (or the wordless <...> placeholder)" },
-	{ "q_dlc_revelation_henry_chapel_common", "Amen." },
-	{ "q_returnToSkalitz_deadPeople", "Jesus..." },
-	{ "event_chase_thiefDown", "Had enough?" },
 	{ "revelation_murderer_ohfuck", "Oh fuck!" },
-	{ "q_dlc_revelation_gambler_diceFail", "Oh, shit!" },
-	{ "q_rides_traitor_kubajsUntied", "Jesus Christ!" },
-	{ "q_returnToSkalitz_deadHangman", "Oh, God." },
-	{ "bowlNotReturned", "What the...?" },
-	{ "q_charlatan_petrDialog", "My condolences." },
 	{ "q_rides_traitor_henry_trail_skirt", "Shameless hussy!" },
-	{ "q_raubritter_playerBloodTrack", "That looks like blood." },
-	{ "revelation_murderer_smell", "Jesus, something stinks here!" },
-	{ "q_dlc_newhomes_judgement_butcher", "What is it this time?" },
-	{ "q_gallowsBrothers_endQuestBark", "So many plots and intrigues..." },
 	{ "q_counterfeiters_crimeScene_area", "Good God, what a bloody mess." },
-	{ "q_theresa_henryLevel_questResult", "I don't know what to say." },
-	{ "q_dlc_revelation_gambler_round2", "Jesus! How did I manage that?" },
-	{ "q_rides_bacchus_ringIsStolenFromHenry", "And what the fuck is this?" },
-	{ "revelation_murderer_trigger_caveBody4", "Jesus Christ, he was only a boy." },
+	{ "event_chase_thiefDown", "Had enough?" },
+	{ "q_returnToSkalitz_deadPeople", "Jesus..." },
+	{ "q_returnToSkalitz_deadHangman", "Oh, God." },
 }
 
 local m = HorseCollisionMod